reconstruct package structure
