--- src/main/java/io/yeahx4/args/NullArgsException.java
@@ -1,7 +1,0 @@
-package io.yeahx4.args;
-
-public final class NullArgsException extends RuntimeException {
-    public NullArgsException(String key) {
-        super("The value of the key " + key + " is null.");
-    }
-}
