+++ src/main/java/io/yeahx4/args/NullArgsValueException.java
@@ -1,0 +1,7 @@
+package io.yeahx4.args;
+
+public final class NullArgsValueException extends RuntimeException {
+    public NullArgsValueException(String key) {
+        super("The value of the key " + key + " is null.");
+    }
+}
