+++ src/main/java/io/yeahx4/args/data/ArgsData.java
@@ -1,4 +1,4 @@
-package io.yeahx4.args;
+package io.yeahx4.args.data;
 
 public interface ArgsData {
 }
