+++ src/main/java/io/yeahx4/args/data/ArgsHolder.java
@@ -1,0 +1,4 @@
+package io.yeahx4.args.data;
+
+public final class ArgsHolder {
+}
