+++ src/main/java/io/yeahx4/args/parser/ArgsParser.java
@@ -1,20 +1,21 @@
 package io.yeahx4.args.parser;
 
-import io.yeahx4.args.ArgsData;
-import io.yeahx4.args.ArgsPair;
-import io.yeahx4.args.ArgsValue;
-import io.yeahx4.args.NullArgsException;
+import io.yeahx4.args.data.ArgsData;
+import io.yeahx4.args.data.ArgsPair;
+import io.yeahx4.args.data.ArgsValue;
+import io.yeahx4.args.NullArgsValueException;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public final class ArgsParser {
     public static List<ArgsData> parse(String[] args) {
-        List<ArgsData> list = new java.util.ArrayList<>(List.of());
+        List<ArgsData> list = new ArrayList<>(List.of());
 
         for (int i = 0; i < args.length; i++) {
             if (args[i].startsWith("--")) {
                 if (i + 1 >= args.length) {
-                    throw new NullArgsException(args[i].substring(2));
+                    throw new NullArgsValueException(args[i].substring(2));
                 }
 
                 list.add(new ArgsPair<>(args[i].substring(2), new ArgsValue<>(args[++i])));
@@ -24,5 +25,9 @@ public final class ArgsParser {
         }
 
         return list;
+    }
+
+    private ArgsParser() {
+        throw new UnsupportedOperationException("This class cannot be instantiated");
     }
 }
